requirements.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
alembic==1.11.3
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asteroid-filterbanks==0.4.0
asttokens==2.2.1
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.0
Babel==2.12.1
backcall==0.2.0
backports.cached-property==1.0.2
beautifulsoup4==4.12.2
bleach==6.0.0
blis==0.7.10
brotlipy==0.7.0
cachetools==5.3.1
catalogue==2.0.9
certifi==2023.7.22
cffi @ file:///C:/b/abs_49n3v2hyhr/croot/cffi_1670423218144/work
chardet==5.2.0
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.7
cmaes==0.10.0
colorama==0.4.6
colorlog==6.7.0
comm==0.1.4
confection==0.1.1
contourpy==1.1.0
cryptography @ file:///C:/b/abs_13590mi9q9/croot/cryptography_1689373706078/work
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.7.post1
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.4
django-cors-headers==4.2.0
djangorestframework==3.14.0
docopt==0.6.2
einops==0.3.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.6.0/en_core_web_lg-3.6.0-py3-none-any.whl#sha256=9012584a8fcf5bc61bedb13a4b511fb81ffff9f770958ee06366671be10632e9
es-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/es_core_news_lg-3.6.0/es_core_news_lg-3.6.0-py3-none-any.whl#sha256=882bd0af2a5117acc17812e821133d8884cfaf4c93362a9ebac538548729de8d
exceptiongroup==1.1.3
executing==1.2.0
fastjsonschema==2.18.0
ffmpeg==1.4
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.6.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.57.0
h5py==3.9.0
hmmlearn==0.2.8
huggingface-hub==0.16.4
HyperPyYAML==1.2.1
idna @ file:///C:/b/abs_bdhbebrioa/croot/idna_1666125572046/work
ipykernel==6.25.1
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.1.0
isoduration==20.11.0
jedi==0.19.0
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
julius==0.2.7
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.2
jupyter_server_terminals==0.4.4
jupyterlab==4.0.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
jupyterlab_server==2.24.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
langcodes==3.3.0
language-tool-python==2.7.1
lazy_loader==0.3
libclang==16.0.6
librosa==0.9.2
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe @ file:///C:/ci/markupsafe_1654508036328/work
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl-fft==1.3.6
mkl-random @ file:///C:/Users/dev-admin/mkl/mkl_random_1682977971003/work
mkl-service==2.4.0
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
msgpack==1.0.5
multidict==6.0.4
murmurhash==1.0.9
nbclient==0.8.0
nbconvert==7.7.4
nbformat==5.9.2
nest-asyncio==1.5.7
networkx==2.8.8
nl-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/nl_core_news_lg-3.6.0/nl_core_news_lg-3.6.0-py3-none-any.whl#sha256=307ca2f66f3e585df5be13e05bd2d6bdb43c3e9e0c33ac75ac1e4e395249b1e1
nltk==3.8.1
notebook==7.0.2
notebook_shim==0.2.3
numba==0.57.1
numpy==1.24.3
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optuna==3.3.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.2
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.10.0
pooch==1.7.0
preshed==3.0.8
primePy==1.3
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.19.6
psutil==5.9.5
psycopg2==2.9.7
pure-eval==0.2.2
pyannote.audio==2.1.1
pyannote.core==4.5
pyannote.database==4.1.3
pyannote.metrics==3.2.1
pyannote.pipeline==2.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==2.3.0
pydantic_core==2.6.3
pyDeprecate==0.3.2
pydub==0.25.1
Pygments==2.16.1
pyOpenSSL @ file:///C:/b/abs_08f38zyck4/croot/pyopenssl_1690225407403/work
pyparsing==3.0.9
pyphen==0.14.0
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.6.5
pytorch-metric-learning==1.7.3
pytz==2023.3
pywin32==306
pywinpty==2.0.11
PyYAML==6.0.1
pyzmq==25.1.1
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.8.8
requests @ file:///C:/b/abs_316c2inijk/croot/requests_1690400295842/work
requests-oauthlib==1.3.1
resampy==0.4.2
Resemblyzer==0.1.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.2
semver==2.13.0
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shellingham==1.5.3
simplejson==3.19.1
singledispatchmethod==1.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
soupsieve==2.4.1
soxr==0.3.6
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
spectralcluster==0.2.20
speechbrain==0.5.15
SQLAlchemy==2.0.20
sqlparse==0.4.4
srsly==2.4.7
stack-data==0.6.2
sympy @ file:///C:/b/abs_95fbf1z7n6/croot/sympy_1668202411612/work
tabulate==0.9.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-hub==0.14.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
textblob==0.17.1
textstat==0.7.3
thinc==8.1.12
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.13.3
tomli==2.0.1
torch==2.0.1+cu117
torch-audiomentations==0.11.0
torch-pitch-shift==1.2.4
torchaudio==2.0.2+cu117
torchmetrics==0.11.4
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.9.0
transformers==4.31.0
typer==0.9.0
typing==3.7.4.3
typing_extensions @ file:///C:/b/abs_213vg2cd1l/croot/typing_extensions_1690297804941/work
tzdata==2023.3
uri-template==1.3.0
urllib3 @ file:///C:/b/abs_889_loyqv4/croot/urllib3_1686163174463/work
wasabi==1.1.2
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.6.2
Werkzeug==2.3.7
widgetsnbextension==4.0.8
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
wrapt==1.15.0
yarl==1.9.2
